[MISC] mm: Extend the gran allocator to handle >32 grandule allocations

+++ mm/mm_gran/mm_gran.h
@@ -80,7 +80,7 @@
 struct gran_s
 {
   uint8_t    log2gran;  /* Log base 2 of the size of one granule */
-  uint16_t   ngranules; /* The total number of (aligned) granules in the heap */
+  uint32_t   ngranules; /* The total number of (aligned) granules in the heap */
 #ifdef CONFIG_GRAN_INTR
   irqstate_t irqstate;  /* For exclusive access to the GAT */
 #else
